snap: support nvidia-gpu-operator in strict confinement
This commit allows microk8s to support the nvidia-gpu-operator when
running in strict confinement, both on core and on classic.

It assumes that the drivers are pre-installed on the host, and that the
runtime libraries are provided via the graphics-core22 slot. The
operator should then be installed with driver.enabled=false and
toolkit.enabled=false.

Signed-off-by: Alex Lewontin <alex.lewontin@canonical.com>